#12 refactor/ Callback interfaceをモジュールに分離
