+++ app/src/main/java/com/kabos/spotifydj/ui/adapter/TrackAdapter.kt
@@ -10,8 +10,9 @@ import com.bumptech.glide.Glide
 import com.kabos.spotifydj.R
 import com.kabos.spotifydj.databinding.ListItemTrackBinding
 import com.kabos.spotifydj.model.TrackInfo
+import com.kabos.spotifydj.util.callback.TrackCallback
 
-class TrackAdapter(private val callback: AdapterCallback)
+class TrackAdapter(private val callback: TrackCallback)
     : ListAdapter<TrackInfo, TrackAdapter.TrackViewHolder>(DiffCallback) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
         val layoutInflater = LayoutInflater.from(parent.context)
