+++ app/src/main/java/com/kabos/spotifydj/util/callback/PlaylistCallback.kt
@@ -1,0 +1,8 @@
+package com.kabos.spotifydj.util.callback
+
+import com.kabos.spotifydj.model.playlist.PlaylistItem
+
+interface PlaylistCallback {
+    fun onClick(playlistItem: PlaylistItem)
+}
+
